Clean up + optimize query
